python 3; pytorch 1.2

+++ src/ScalingAndSquaring.py
@@ -12,7 +12,7 @@ class ScalingAndSquaring(nn.Module):
     x = x/(2**self.num_steps)
     for _ in range(self.num_steps):
       xDef = torch.empty(x.shape, device=x.device, requires_grad=False)
-      for chanIdx in range(-1, (x.shape[1]/3) - 1):
+      for chanIdx in range(-1, int(x.shape[1]/3) - 1):
         chanRange = range(chanIdx * 3, chanIdx * 3 + 3)
         for channel in chanRange:
           imgToDef = x[:, None, channel, ]
